+++ packages/ui-components/src/components/ui/combobox-field/combobox-field.tsx
@@ -37,6 +37,7 @@ export interface ComboboxFieldProps<OptionType>
   disabled?: boolean;
   value?: string | null;
   onChange?: (value: string | null) => void;
+  inputValue?: string;
   onInputValueChange?: (value: string) => void;
 }
 
@@ -52,6 +53,7 @@ function ComboboxField<OptionType>({
   options,
   renderItemLabel,
   onChange,
+  inputValue,
   onInputValueChange,
   getOptionLabel = (val) => (val as { label: string }).label,
   getOptionValue = (val) => (val as { value: string | null }).value,
@@ -74,6 +76,7 @@ function ComboboxField<OptionType>({
         onValueChange={(option) => {
           onChange?.(option ? getOptionValue(option) : null);
         }}
+        inputValue={inputValue}
         onInputValueChange={onInputValueChange}
         disabled={disabled}
         items={options}