+++ packages/core-generators/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @baseplate-dev/core-generators
 
+## 0.6.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @baseplate-dev/sync@0.6.4
+  - @baseplate-dev/utils@0.6.4
+
 ## 0.6.3
 
 ### Patch Changes