+++ packages/sync/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @baseplate-dev/sync
 
+## 0.6.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @baseplate-dev/utils@0.6.4
+
 ## 0.6.3
 
 ### Patch Changes